Update AESTubeWav.py
Right so, this update fixed an issue that I forgot to test for and that's the decrypt process where you could use a different YouTube link to decrypt the file. youtube-dl will auto resume if it finds a file by a matching name. I could've done a couple of things to fix this. However, this bug was fixed with the update. 
Now the file gets saved to mediafiles/ and once the program is completed it gets deleted. The audio files are now rearranged for stronger encryption.
I added all 88 piano keys at the A4(440HZ) to the freq_array. 
Added two bash scripts for rearranging the audio file.